Test structure cleanup
